feat: consolidate projects-add-one and projects-update-one into projects-manage
- Merges projects-add-one and projects-update-one into a single projects-manage tool
- Uses natural parameter detection: if id is provided → update, else → add
- For updates: requires id and name
- For creates: requires only name
- Maintains mapProject for new projects, preserves original return type for updates
- Reduces tool count from 14 to 13 tools
